refactor: combat phase-2 slice 11 — general assignment and mode tables (#3865)
Migrate battle_general_assignment_*, bind phase, and combat_mode_selection
suites to table-driven scenario loops; combat test LOC 1921 → 1580.

Refs #3865
